refactor: restructure utils/ into domain subdirectories (formatters, validation, bot)
Move 8 files from utils/ root into thematic subdirectories for consistency
with handlers/ structure. Update all imports across 20+ consumer files.
