+++ handlers/tests/test_taking.py
@@ -27,7 +27,7 @@ from keyboards.keyboards import get_simple_test_selection_keyboard, get_test_sta
 from aiogram.types import InlineKeyboardMarkup, InlineKeyboardButton
 from states.states import TestTakingStates
 from utils.logger import log_user_action, log_user_error, logger
-from utils.test_progress_formatters import get_test_status_icon
+from utils.formatters.test_progress import get_test_status_icon
 from utils.auth.auth import check_auth
 from utils.handlers.user import get_current_user
 from utils.handlers.callback import ensure_callback_auth
