+++ tests/unit/utils/test_test_progress_formatters.py
@@ -1,6 +1,6 @@
 """Тесты для модуля форматирования прогресса тестов"""
 import pytest
-from utils.test_progress_formatters import (
+from utils.formatters.test_progress import (
     get_test_status_icon,
     format_test_line,
     format_test_with_percentage,
