+++ utils/bot/errors.py
